mcp/azure-mcp/requirements.txt
mcp[cli]
azure-identity
openai
python-dotenv
prettytable
websockets
pandas